Move popup scrim to class
